Suggest mode: relocate wordDiff to its own module
Extract the word-level diff (wordDiff plus its private tokenize and LCS
helpers) out of the overlay-era suggestion-diff.js into a standalone
word-diff.js. suggestion-summary.js is wordDiff's only surviving consumer;
giving it a dedicated home lets the overlay render component suggestion-diff.js
be deleted with the rest of the overlay in a later step without taking wordDiff
down with it. Pure move: no behavior change, tests relocated alongside.
