+++ packages/editor/src/components/suggestion-mode/index.js
@@ -44,7 +44,8 @@ export {
 	PAYLOAD_MAX_BYTES,
 	SCHEMA_VERSION,
 } from './provider';
-export { default as SuggestionDiff, wordDiff } from './suggestion-diff';
+export { default as SuggestionDiff } from './suggestion-diff';
+export { wordDiff } from './word-diff';
 export {
 	default as SuggestionSummary,
 	summarizeOperations,
